add TEST_ASSERTS handling; fix classpath; auto-prefix perf improvements

+++ src/python/autoPrefixPerf.py
@@ -6,7 +6,7 @@ import re
 # run from lucene subdir in trunk checkout
 
 reIter = re.compile(': (\d+) msec; totalHits=(\d+) hash=(\d+)')
-logsDir = '/x/tmp/prefixtermsperf'
+logsDir = '/x/tmp/prefixtermsperf2'
 
 def run(cmd):
   if os.system(cmd):
@@ -59,22 +59,24 @@ run('javac -d /l/util/build -cp build/core/classes/java:build/analysis/common/cl
 totalHits = None
 hash = None
 
-for precStep in (4, 8, 12, 16):
-  print
-  print('NF precStep=%d' % precStep)
-  logFileName = '%s/nf.precStep%d.txt' % (logsDir, precStep)
-  if not os.path.exists(logFileName):
-    if os.path.exists('/l/indices/numbers'):
-      shutil.rmtree('/l/indices/numbers')
-    run('java -cp /l/util/build:build/core/classes/java:build/analysis/common/classes/java perf.AutoPrefixPerf /lucenedata/numbers/randlongs.10m.txt /lucenedata/numbers/randlongs.queries.txt /l/indices/numbers %d 0 0 > %s 2>&1' % (precStep, logFileName))
+if False:
+  for precStep in (4, 8, 12, 16):
+    print
+    print('NF precStep=%d' % precStep)
+    logFileName = '%s/nf.precStep%d.txt' % (logsDir, precStep)
+    if not os.path.exists(logFileName):
+      if os.path.exists('/l/indices/numbers'):
+        shutil.rmtree('/l/indices/numbers')
+      run('java -cp /l/util/build:build/core/classes/java:build/analysis/common/classes/java perf.AutoPrefixPerf /lucenedata/numbers/randlongs.10m.txt /lucenedata/numbers/randlongs.queries.txt /l/indices/numbers %d 0 0 > %s 2>&1' % (precStep, logFileName))
 
-  indexTimeSec, indexSizeBytes, totalTerms, bestMS = parseLog(logFileName)
-  print('    index sec %.2f' % indexTimeSec)
-  print('    index MB %.2f' % (indexSizeBytes/1024/1024.))
-  print('    term count %s' % totalTerms)
-  print('    search msec %s' % bestMS)  
+    indexTimeSec, indexSizeBytes, totalTerms, bestMS = parseLog(logFileName)
+    print('    index sec %.2f' % indexTimeSec)
+    print('    index MB %.2f' % (indexSizeBytes/1024/1024.))
+    print('    term count %s' % totalTerms)
+    print('    search msec %s' % bestMS)  
 
 for minItemsInPrefix in 5, 10, 15, 20, 25, 30, 35, 40:
+#for minItemsInPrefix in 80, 90, 100, 110, 120:
   for mult in 2, 3, 4, 5, None:
     if mult is None:
       maxItemsInPrefix = (1<<31)-1
